StudyApplicationService 단위 테스트 추가

+++ src/main/java/com/example/studymate/member/repository/MemberRepository.java
@@ -1,10 +1,10 @@
 package com.example.studymate.member.repository;
 
-import com.example.studymate.member.entity.Member;
+import com.example.studymate.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface MemberRepository extends JpaRepository<Member, Long> {
+public interface MemberRepository extends JpaRepository<Member, Integer> {
     Optional<Member> findByEmail(String email);
     
 }
